port: ble: Rename the port API to crypto
All the port layer to support BLE network is actually just
cryptographic APIs that actually depends on a crypto implementation
and not in the RTOs/port.

Rename them to use the namespace crypto and simplify the API
getting rid of the vtable and declaring the functions directly
since none of them are optional, or have callback or intended to
change at runtime.

Signed-off-by: Flavio Ceolin <flavio@hubblenetwork.com>